+++ config/config.go
@@ -1,0 +1,6 @@
+package config
+
+var (
+	TerminalHeight int
+	TerminalWidth  int
+)