+++ layout.go
@@ -2,28 +2,24 @@ package main
 
 import (
 	"github.com/charmbracelet/lipgloss"
+	"github.com/jonnypolite/imbox/config"
 	"github.com/jonnypolite/imbox/style"
 )
 
 const ListBoxHeight int = 10
 
-var (
-	terminalHeight int
-	terminalWidth  int
-)
-
 func ListBox(content string, selected bool) string {
-	return style.BoxStyle(ListBoxHeight, terminalWidth, selected).
+	return style.BoxStyle(ListBoxHeight, config.TerminalWidth, selected).
 		Render(content)
 }
 
 func ReadBoxStyle(selected bool) lipgloss.Style {
-	return style.BoxStyle(readBoxHeight(), terminalWidth, selected)
+	return style.BoxStyle(readBoxHeight(), config.TerminalWidth, selected)
 }
 
 func SetTerminalSize(height int, width int) {
-	terminalHeight = height
-	terminalWidth = width
+	config.TerminalHeight = height
+	config.TerminalWidth = width
 }
 
 func TitleBar(titleText string, width int) string {
@@ -35,5 +31,5 @@ func TitleBar(titleText string, width int) string {
 }
 
 func readBoxHeight() int {
-	return terminalHeight - ListBoxHeight - 5
+	return config.TerminalHeight - ListBoxHeight - 5
 }