+++ main.go
@@ -7,8 +7,11 @@ import (
 	"github.com/charmbracelet/bubbles/viewport"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
+	"github.com/jonnypolite/imbox/config"
 	"github.com/jonnypolite/imbox/mailbox"
+	"github.com/jonnypolite/imbox/style"
 	"github.com/jonnypolite/imbox/ui"
+	"github.com/jonnypolite/imbox/ui/helpers"
 )
 
 var cli struct {
@@ -27,10 +30,12 @@ const (
 const numberOfBoxes = 2
 
 type mainModel struct {
-	selectedBox box
-	emails      []mailbox.Email
-	summaryList ui.ScrollingList[mailbox.MailSummary]
-	bodyView    ui.BodyView
+	selectedBox   box
+	emails        []mailbox.Email
+	summaryList   ui.ScrollingList[mailbox.MailSummary]
+	bodyView      ui.BodyView
+	dialogVisible bool
+	dialog        ui.Confirm
 }
 
 // Init is called just before the first render
@@ -38,35 +43,58 @@ func (m mainModel) Init() tea.Cmd {
 	return nil
 }
 
+func (m *mainModel) showConfirm(confirmText, yesText, noText string) {
+	m.dialog = ui.NewConfirm(confirmText, yesText, noText)
+	m.dialogVisible = true
+}
+
+// Updates the model. Executes every time there is an action
+// like a key press.
 func (m mainModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	switch message := msg.(type) {
 	case tea.WindowSizeMsg:
 		SetTerminalSize(message.Height, message.Width)
 
 	case tea.KeyMsg:
-		switch message.String() {
-		case "ctrl+c", "q":
-			return m, tea.Quit
-		case "tab":
-			m.selectedBox = (m.selectedBox + 1) % numberOfBoxes
-		case "j":
-			if m.selectedBox == listBox {
-				m.bodyView.SetEmail(m.emails[m.summaryList.Down()])
+		if m.dialogVisible {
+			selectionMade, selection := m.dialog.Update(msg)
+			if selectionMade {
+				if selection {
+					return m, tea.Quit
+				}
+				m.dialogVisible = false
 			}
-		case "k":
-			if m.selectedBox == listBox {
-				m.bodyView.SetEmail(m.emails[m.summaryList.Up()])
+		} else if m.selectedBox == readBox {
+			switch message.String() {
+			case "ctrl+c", "q":
+				m.showConfirm("Are you sure you want to quit?", "Yes", "No")
+			case "tab":
+				m.selectedBox = (m.selectedBox + 1) % numberOfBoxes
+			default:
+				m.bodyView.Viewport, _ = m.bodyView.Viewport.Update(msg)
+			}
+		} else {
+			switch message.String() {
+			case "ctrl+c", "q":
+				m.showConfirm("Are you sure you want to quit?", "Yes", "No")
+			case "tab":
+				m.selectedBox = (m.selectedBox + 1) % numberOfBoxes
+			case "j":
+				if m.selectedBox == listBox {
+					m.bodyView.SetEmail(m.emails[m.summaryList.Down()])
+				}
+			case "k":
+				if m.selectedBox == listBox {
+					m.bodyView.SetEmail(m.emails[m.summaryList.Up()])
+				}
 			}
 		}
-	}
-
-	if m.selectedBox == readBox {
-		m.bodyView.Viewport, _ = m.bodyView.Viewport.Update(msg)
 	}
 
 	return m, nil
 }
 
+// View builds the complicated string that is the UI.
 func (m mainModel) View() string {
 	var s string
 	readBoxStyle := ReadBoxStyle(m.selectedBox == readBox)
@@ -76,13 +104,22 @@ func (m mainModel) View() string {
 
 	s += lipgloss.JoinVertical(
 		lipgloss.Top,
-		TitleBar("Imbox", terminalWidth),
+		TitleBar("Imbox", config.TerminalWidth),
 		ListBox(
-			m.summaryList.Display(terminalWidth-2),
+			m.summaryList.Display(config.TerminalWidth-2),
 			m.selectedBox == listBox,
 		),
 		readBoxStyle.Render(m.bodyView.View()),
 	)
+
+	if m.dialogVisible {
+		return helpers.PlaceOverlay(
+			(config.TerminalWidth/2)-(style.ConfirmBoxStyle.GetWidth()/2),
+			10,
+			m.dialog.View(),
+			s,
+		)
+	}
 
 	return s
 }