+++ style/colors.go
@@ -1,10 +1,17 @@
 package style
 
 const (
+	// Button Styles
+	ButtonSelectedBG = "#bd93f9"
+	ButtonSelectedFG = "#f8f8f2"
+
+	// List Styles
+	ListSelectedBG = "#bd93f9"
+	ListSelectedFG = "#f8f8f2"
+
+	// Box Styles
 	SelectedBoxBG       = "#282a36"
-	UnselectedBoxBG     = "#000000"
-	ListSelectedBG      = "#bd93f9"
-	ListSelectedFG      = "#f8f8f2"
 	SelectedBoxBorder   = "#50fa7b"
+	UnselectedBoxBG     = "#282a36"
 	UnselectedBoxBorder = "#a9aaab"
 )