+++ style/styles.go
@@ -21,3 +21,21 @@ func BoxStyle(height int, width int, selected bool) lipgloss.Style {
 		Height(height).
 		Width(width - 2) // The -2 accounts for the width of the borders
 }
+
+var ConfirmBoxStyle = lipgloss.NewStyle().
+	BorderStyle(lipgloss.DoubleBorder()).
+	BorderForeground(lipgloss.Color(SelectedBoxBorder)).
+	Height(5).
+	Width(35).
+	Align(lipgloss.Center)
+
+var ButtonStyle = lipgloss.NewStyle().
+	Foreground(lipgloss.Color("#FFF7DB")).
+	Background(lipgloss.Color("#888B7E")).
+	Padding(1, 3).
+	Margin(2)
+
+var ActiveButtonStyle = ButtonStyle.
+	Foreground(lipgloss.Color(ButtonSelectedFG)).
+	Background(lipgloss.Color(ButtonSelectedBG)).
+	Underline(true)